fix - improvment of error messages

+++ src/services/users/delete-user.ts
@@ -1,6 +1,7 @@
 import { CheckInsRepository } from '@/repositories/check-ins-repository'
 import { UsersRepository } from '@/repositories/users-repository'
 import { CannotDeleteUserError } from '../errors/cannot-delete-user'
+import { UserNotFoundError } from '../errors/user-not-found'
 
 export class DeleteService {
   constructor(
@@ -10,9 +11,13 @@ export class DeleteService {
 
   async execute(userId: string): Promise<boolean> {
     const checkIn = await this.checkInRepository.findByUserId(userId)
-    console.log(checkIn)
+    const user = await this.usersRepository.findByUserId(userId)
+
     if (checkIn?.id) {
       throw new CannotDeleteUserError()
+    }
+    if (!user) {
+      throw new UserNotFoundError()
     }
     const success = await this.usersRepository.delete(userId)
